+++ packages/sapa/src/functions/vnode.js
@@ -270,7 +270,15 @@ export class VNode {
             }
           });
         } else if (isFunction(child)) {
-          fragment.appendChild(child());
+          const result = child();
+
+          if (result instanceof VNode) {
+            const el = result.makeElement(withChildren, options);
+
+            if (el) fragment.appendChild(el);
+          } else if (typeof result === "string") {
+            fragment.appendChild(document.createTextNode(result));
+          }
         } else if (child instanceof window.HTMLElement) {
           fragment.appendChild(child);
         } else {
@@ -285,12 +293,14 @@ export class VNode {
       children.forEach((child) => {
         if (isArray(child)) {
           child.forEach((it) => {
-            if (it) {
+            if (isFunction(it?.runMounted)) {
               it.runMounted();
             }
           });
         } else if (child) {
-          child.runMounted();
+          if (isFunction(child?.runMounted)) {
+            child.runMounted();
+          }
         }
       });
     }