+++ packages/sapa/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @elf-framework/sapa
 
+## 2.0.20
+
+### Patch Changes
+
+- fixed vnode mounted hook bug
+
+## 2.0.16
+
+### Patch Changes
+
+- fixed vnode function element
+
 ## 2.0.12
 
 ### Patch Changes